# perf: use only setImmediate