@pi0 2021-04-09
Hi @ChangJoo-Park. Thank you for the benchmarks and PR. But we need a promise wrapper. Since using setImmediate only, we were resolving h3 handle before the response is sent. It was breaking nitro since the response was not ready after resolving (we may workaround for nitro but I think in-between perf and stability of promise+immediate is good enough for now...)