@ChangJoo-Park 2021-04-09
Thank you for reply @pi0! 😄 I didn't know nitro needed Promise. Thanks for letting me know. I'm waiting for nitro to be open sourced. It is too hard to read the built nitro source 😭