@pi0 2021-04-09
More precisely promise resolution was required for mocked version of HTTP module for direct ssr calls. We are hardly trying to make opensourcing it as soon as possible. BTW the related mock if interested to read: https://unpkg.com/browse/@nuxt/un@0.1.1/runtime/fetch/call.js